refactor: css 이름 변경 및 SectorApplyModal로 분리
